+++ README.md
@@ -1,2 +1,23 @@
 # Artemis-dev-tool
 
+Project Artemis is an Open Sourced developer tool for testing GraphQL queries for projects that use Apollo Client
+
+Core features:
+1. Query caching and component storage
+2. 
+3. Performance monitoring and Management
+4. Query response Snapshot feature
+
+Technologies used: 
+React Hooks, Typescript, GraphQL, 
+Jest, Enzyme, and Testing
+
+
+End to end query management
+and query facilitation 
+Simply go to the chrome web store and download:
+
+
+
+
+npm package:
