+++ package.json
@@ -20,7 +20,18 @@
   },
   "homepage": "https://github.com/timjj92/Artemis-dev-tool#readme",
   "dependencies": {
+    "@material-ui/core": "^4.7.2",
+    "@material-ui/icons": "^4.5.1",
+    "@types/enzyme": "^3.10.4",
+    "@types/jest": "^24.0.23",
+    "@types/mocha": "^5.2.7",
+    "@types/react": "^16.9.16",
+    "@types/react-dom": "^16.9.4",
+    "awesome-typescript-loader": "^5.2.1",
+    "enzyme": "^3.10.0",
     "graphiql": "^0.17.0",
+    "jest": "^24.9.0",
+    "mocha": "^6.2.2",
     "react": "^16.12.0",
     "react-bootstrap": "^1.0.0-beta.16",
     "react-d3-tree": "^1.15.3",
@@ -29,23 +40,28 @@
     "react-json-view": "^1.19.1",
     "react-live-clock": "^3.1.0",
     "semantic-ui-react": "^0.88.1",
+    "source-map-loader": "^0.2.4",
+    "ts-node": "^8.5.4",
     "victory": "^33.1.7"
   },
   "devDependencies": {
     "@babel/core": "^7.7.4",
-    "@babel/preset-env": "^7.7.4",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/preset-env": "^7.7.4",
     "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.7.4",
+    "@types/chrome": "0.0.91",
     "babel-loader": "^8.0.6",
-    "html-webpack-plugin": "^3.2.0",
     "concurrently": "^5.0.0",
     "css-loader": "^3.2.0",
+    "html-webpack-plugin": "^3.2.0",
     "node-sass": "^4.13.0",
     "sass-loader": "^8.0.0",
     "style-loader": "^1.0.0",
     "terser-webpack-plugin": "^1.2.3",
+    "ts-loader": "^6.2.1",
+    "typescript": "^3.7.3",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.10",
     "webpack-dev-server": "^3.9.0"
