+++ src/containers/ObserverContainer.tsx
@@ -1,4 +1,16 @@
 import React from "react";
+
+import { Theme, createStyles, makeStyles } from '@material-ui/core/styles';
+import ExpansionPanel from '@material-ui/core/ExpansionPanel';
+import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
+import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
+import Typography from '@material-ui/core/Typography';
+import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
+import RemoveIcon from '@material-ui/icons/Remove';
+import IconButton from '@material-ui/core/IconButton';
+import Button from '@material-ui/core/Button';
+
+
 import HistoryOfPastQueries from "../components/HistoryOfPastQueries.jsx";
 import Query from "../components/Query.jsx";
 import GraphQLResponse from "../components/GraphQLResponse.jsx";
@@ -8,8 +20,36 @@ import GraphQLVisualizer from '../components/GraphQLVisualizer.jsx';
 
 import ChartComponent from "../components/ChartComponent.jsx";
 
+/**
+ * ************************************
+ *
+ * @module  ObserverCOntainers
+ * @description
+ *
+ * ************************************
+ */
+const useStyles = makeStyles((theme: Theme) => createStyles({
+  root: {
+    width: '100%',
+  },
+  heading: {
+    fontSize: theme.typography.pxToRem(15),
+    fontWeight: theme.typography.fontWeightRegular,
+  },
+  button: {
+    marginLeft: '5px',
+  },
+}));
 
-const ObserverContainers = props => {
+interface Props {
+  name: string,
+  attributes?: {id: string}
+
+}
+
+const ObserverContainers = (props)=> {
+
+  const classes = useStyles({});
 
   return (
     <React.Fragment>
