+++ webpack.config.js
@@ -1,5 +1,4 @@
 const path = require('path');
-
 module.exports = {
   entry: path.join(__dirname, 'src/panel.js'), //where to build dependency graph 
   output: {
@@ -10,13 +9,6 @@ module.exports = {
   module:{
     rules: [
       {
-        test: /\.tsx?$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'ts-loader'
-        }
-      },
-      {
         test: /\.jsx?/,//if file ends with this text run this 
         exclude: /(node_modules|bower_components)/,
         use: {
@@ -24,6 +16,13 @@ module.exports = {
           options: {
             presets: ['@babel/preset-env','@babel/preset-react']
           }
+        }
+      },
+      {
+        test: /\.tsx?$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'ts-loader'
         }
       },
       {
@@ -44,3 +43,7 @@ module.exports = {
   },
   devtool: 'inline-source-map'
 };
+
+
+
+
